major updates
nav_sys_flight is the most current control algorithm for the quadcopter.
It utilizes the compass and gps in conjunction with the law of cosines
to determine the angle of rotation necesarry to be facing the target gps
coord. It also uses the distance formula to determine the distance
between the current position and the target. It has sd card storage
capabilities as well as the ability to arm the quad copter esc's and
begin controlling the motors.